Feat: rotation is happening successfully,test passed

+++ src/main/java/com/yourorg/logging/config/LoggerConfig.java
@@ -43,7 +43,7 @@ public class LoggerConfig {
     }
 
     public static class Retention {
-        public int rotateSizeMB;
+        public double rotateSizeMB;
         public int maxDays;
     }
 }
